requirements.txt
fastapi
uvicorn
pydantic
requests
kafka-python
hazelcast-python-client
python-consul
